+++ hpx/performance_counters/registry.hpp
@@ -166,12 +166,6 @@ namespace hpx { namespace performance_counters
     private:
         counter_type_map_type countertypes_;
     };
-
-    namespace detail
-    {
-        HPX_EXPORT std::string regex_from_pattern(std::string const& pattern,
-            error_code& ec);
-    }
 }}
 
 #endif
